+++ CHANGELOG.md
@@ -9,7 +9,9 @@
   - property `wrongMethodBehavior` (number) changed to `hintAllowedMethods` (boolean);
   - property `methodLikeRouteBehavior` (string literal) changed to `recognizeMethodDependentRoutes` (boolean);
 - Breaking change to the `Documentation` constructor argument (object):
-  - property `hasSummaryFromDescription` (boolean) renamed to `hasSummary` (boolean);
+  - property `hasSummaryFromDescription` renamed to `hasSummary`;
+- Breaking change to the `Integration` constructor argument (object):
+  - property `noContent` renamed to `noBodySchema`;
 - Consider using [the automated migration](https://www.npmjs.com/package/@express-zod-api/migration).
 
 ```diff
@@ -25,6 +27,13 @@
   new Documentation({
 -   hasSummaryFromDescription: false,
 +   hasSummary: false,
+  });
+```
+
+```diff
+  new Integration({
+-   noContent: z.undefined(),
++   noBodySchema: z.undefined(),
   });
 ```